Trajectory computations as separate class/file.
